refactor(components): remove redunant props + make correct docs
