+++ src/components/Switch/Switch.tsx
@@ -12,15 +12,14 @@ type CustomCSSProperties = React.CSSProperties & {
 };
 
 export type SwitchProps = {
+  /** Optional label for the switch */
   label?: string;
+  /** Optional size of the switch */
   size?: CheckboxSize;
-  color?: string;
-  type?: 'checkbox';
-  className?: string;
 } & Omit<React.InputHTMLAttributes<HTMLInputElement>, 'size'>;
 
-// A functional component that renders a customizable switch input with an optional label.
-// This component supports various sizes and colors, and can be disabled or checked by default.
+/** A functional component that renders a customizable switch input with an optional label.
+ * This component supports various sizes and colors, and can be disabled or checked by default. */
 export const Switch: FC<SwitchProps> = ({
   label,
   size = 'medium',
